requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
pandas>=2.2.0
openpyxl>=3.1.2